Move `MacroExpansion{Error|Warning|FixIt}Message` to the `SwiftSyntaxMacros` module
The expansion diagnostic messages were defined in `SwiftSyntaxMacroExpansion`, which is intended as an implementation detail of the plugin server and should not need to be imported by macros. Move them to the `SwiftSyntaxMacros` module where they belong.
