+++ Sources/SwiftSyntaxMacros/CMakeLists.txt
@@ -22,6 +22,7 @@ add_swift_syntax_library(SwiftSyntaxMacros
 
   AbstractSourceLocation.swift
   MacroExpansionContext.swift
+  MacroExpansionDiagnosticMessages.swift
 )
 
 target_link_swift_syntax_libraries(SwiftSyntaxMacros PUBLIC
